Dynamic master node base database
