+++ alpha/alpha-server/src/main/java/org/apache/servicecomb/pack/alpha/server/cluster/master/provider/jdbc/JdbcLockProvider.java
@@ -15,15 +15,13 @@
  * limitations under the License.
  */
 
-package org.apache.servicecomb.pack.alpha.server.cluster.lock.provider;
+package org.apache.servicecomb.pack.alpha.server.cluster.master.provider.jdbc;
 
-import org.apache.servicecomb.pack.alpha.server.cluster.lock.LockConfig;
+import org.apache.servicecomb.pack.alpha.server.cluster.master.provider.AbstractLockProvider;
+import org.apache.servicecomb.pack.alpha.server.cluster.master.provider.jdbc.jpa.MasterLockRepository;
 
-public interface LockProviderPersistence {
-
-    boolean initLock(LockConfig lockConfig);
-
-    boolean updateLock(LockConfig lockConfig);
-
-    void unLock(LockConfig lockConfig);
+public class JdbcLockProvider extends AbstractLockProvider {
+    public JdbcLockProvider(MasterLockRepository masterLockRepository) {
+        super(new JdbcLockPersistence(masterLockRepository));
+    }
 }
