+++ alpha/alpha-server/src/main/java/org/apache/servicecomb/pack/alpha/server/cluster/master/provider/jdbc/LockProviderJdbcConfiguration.java
@@ -15,10 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.servicecomb.pack.alpha.server.cluster.lock.provider.jdbc;
+package org.apache.servicecomb.pack.alpha.server.cluster.master.provider.jdbc;
 
-import org.apache.servicecomb.pack.alpha.server.cluster.lock.provider.LockProvider;
-import org.apache.servicecomb.pack.alpha.server.cluster.lock.provider.jdbc.jpa.MasterLockRepository;
+import org.apache.servicecomb.pack.alpha.server.cluster.master.provider.LockProvider;
+import org.apache.servicecomb.pack.alpha.server.cluster.master.provider.jdbc.jpa.MasterLockRepository;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
